rust/gigasecond: simplify typing

+++ rust/gigasecond/src/lib.rs
@@ -2,5 +2,5 @@ use chrono::{DateTime, Duration, Utc};
 
 // Returns a Utc DateTime one billion seconds after start.
 pub fn after(start: DateTime<Utc>) -> DateTime<Utc> {
-    start + Duration::seconds(10u64.pow(9) as i64)
+    start + Duration::seconds(10i64.pow(9))
 }
